Move image into its own component to avoid a renderSomething() function
